+++ NEWS.md
@@ -9,5 +9,11 @@ output: html_document
 knitr::opts_chunk$set(echo = TRUE)
 ```
 
+## 1.0.1
+8th July 2020
+
+* split `Medication.R` to `Medication.R`, `methods.R` and `userInterface.R`
+* change license to Mozilla Public License 2.0
+
 ## 1.0.0
 1st July 2020